+++ src/Curiosity.Migrations/DbMigrator.cs
@@ -112,10 +112,12 @@ namespace Curiosity.Migrations
                     _logger?.LogInformation($"Creating {_dbProvider.MigrationHistoryTableName} table...");
                     await _dbProvider.CreateHistoryTableIfNotExistsAsync(token);
                     _logger?.LogInformation($"Creating {_dbProvider.MigrationHistoryTableName} table completed.");
-                }                
+                }
+
+                var isDowngradeEnabled = _downgradePolicy != MigrationPolicy.Forbidden;
                 
                 _logger?.LogInformation("Check database version...");
-                var dbVersion = await _dbProvider.GetDbVersionSafeAsync(token) ?? default;
+                var dbVersion = await _dbProvider.GetDbVersionSafeAsync(isDowngradeEnabled, token) ?? default;
 
                 var targetVersion = _targetVersion ?? _migrations.Max(x => x.Version);
                 if (targetVersion == dbVersion)
@@ -140,7 +142,7 @@ namespace Curiosity.Migrations
 
                     // DB version might be changed after pre-migration
                     _logger?.LogInformation("Check database version after pre migration...");
-                    dbVersion = await _dbProvider.GetDbVersionSafeAsync(token) ?? default;   
+                    dbVersion = await _dbProvider.GetDbVersionSafeAsync(isDowngradeEnabled, token) ?? default;   
                     _logger?.LogInformation("Check database version after pre migration...");
                 }
                 else
@@ -227,6 +229,9 @@ namespace Curiosity.Migrations
         /// <exception cref="MigrationException"></exception>
         private async Task UpgradeAsync(DbVersion actualVersion, DbVersion targetVersion, CancellationToken token = default)
         {
+            if (_upgradePolicy == MigrationPolicy.Forbidden)
+                throw new MigrationException(MigrationError.PolicyError, "Upgrading is forbidden due to migration upgrade policy");
+            
             var desiredMigrations = _migrations
                 .Where(x => x.Version > actualVersion && x.Version <= targetVersion)
                 .OrderBy(x => x.Version)
@@ -311,12 +316,19 @@ namespace Curiosity.Migrations
         /// <exception cref="MigrationException"></exception>
         private async Task DowngradeAsync(DbVersion actualVersion, DbVersion targetVersion, CancellationToken token = default)
         {
+            if (_downgradePolicy == MigrationPolicy.Forbidden)
+                throw new MigrationException(MigrationError.PolicyError, "Downgrading is forbidden due to migration downrgade policy");
+            
             var desiredMigrations = _migrations
                 .Where(x => x.Version <= actualVersion && x.Version >= targetVersion)
                 .OrderByDescending(x => x.Version)
                 .ToList();
             if (desiredMigrations.Count == 0 || desiredMigrations.Count == 1) return;
 
+            var downgradableMigrationsCount = _migrations.Count(x => x is IDowngradeMigration);
+            if (downgradableMigrationsCount != desiredMigrations.Count)
+                throw new MigrationException(MigrationError.MigrationNotFound, $"Found {downgradableMigrationsCount} downgrade migrations but expected {desiredMigrations.Count}");
+            
             var lastMigrationVersion = new DbVersion(0, 0);
             var currentDbVersion = actualVersion;
             for (var i = 0; i < desiredMigrations.Count - 1; ++i)
@@ -325,6 +337,9 @@ namespace Curiosity.Migrations
                 
                 var targetLocalVersion = desiredMigrations[i + 1].Version;
                 var migration = desiredMigrations[i];
+                
+                if (!(migration is IDowngradeMigration downgradeMigration))
+                    throw new MigrationException(MigrationError.MigrationNotFound, $"Migration {migration.Version} doesn't support downgrade");
 
                 if (!IsMigrationAllowed(DbVersion.GetDifference(currentDbVersion, targetLocalVersion),
                     _downgradePolicy))
@@ -344,8 +359,8 @@ namespace Curiosity.Migrations
                     {
                         _logger?.LogInformation(
                             $"Downgrade to {desiredMigrations[i + 1].Version} (DB {_dbProvider.DbName})...");
-                        await migration.DowngradeAsync(transaction, token);
-                        await _dbProvider.UpdateCurrentDbVersionAsync(migration.Comment, targetLocalVersion, token);
+                        await downgradeMigration.DowngradeAsync(transaction, token);
+                        await _dbProvider.UpdateCurrentDbVersionAsync(downgradeMigration.Comment, targetLocalVersion, token);
                         lastMigrationVersion = targetLocalVersion;
                         currentDbVersion = targetLocalVersion;